feat. 수분 조회 테스트 진행 중

+++ lib/domain/mois.dart
@@ -10,7 +10,7 @@ class Mois {
 
   // 수분량
   @JsonKey(name: 'amount_mois')
-  final int amountMois;
+  final double amountMois;
 
   @JsonKey(name: 'mois_date')
   final DateTime moisDate;
